Move more game-specific modules into their proper subprojects
